% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build.matrix.from.traits.R
\name{build.matrix.from.traits}
\alias{build.matrix.from.traits}
\title{Build a predicted adjacency matrix based on trait values}
\usage{
build.matrix.from.traits(adj, corr.traits)
}
\arguments{
\item{adj}{The adjacency matrix}

\item{corr.traits}{A list with species as rows, the values of the corrected traits as columns, and
the different traits as components}
}
\value{
The function returns a three-dimensional array where the first and second dimension describes
the species, i.e., the adjacency matrix based on the corrected trait values, and the third dimension
corresponds to the trait analyzed.
}
\description{
Build a predicted adjacency matrix based on trait values
}
\examples{
data(eklof)
overlap.matrices <- build.matrix.from.traits(eklof$network, corr.traits)
}
\references{
A Eklöf, U Jacob, J Kopp, J Bosch, R Castro-Urgal, NP Chacoff,
   B Dalsgaard, C de Sassi, M Galetti, PR Guimarães, SB Lomáscolo, AM Martín
   González, MA Pizo, R Rader, A Rodrigo, JM Tylianakis, DP Vázquez, and S
   Allesina (2013) The dimensionality of ecological networks. Ecology
   Letters 16:577-583 (doi: 10.1111/ele.12081)
}

