% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eklof.R
\docType{data}
\name{eklof}
\alias{eklof}
\title{Sample food-web dataset with which to study phylogenetic intervality}
\format{\preformatted{List of 2
 $ network: int [1:69, 1:69] 0 0 0 0 0 0 0 0 0 0 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:69] "sp_1" "sp_2" "sp_3" "sp_4" ...
  .. ..$ : chr [1:69] "sp_1" "sp_2" "sp_3" "sp_4" ...
 $ tree   :List of 5
  ..$ edge       : int [1:104, 1:2] 70 71 72 72 73 74 74 74 75 75 ...
  ..$ Nnode      : int 36
  ..$ tip.label  : chr [1:69] "sp_1" "sp_2" "sp_3" "sp_9" ...
  ..$ edge.length: num [1:104] 0.0147 0.4118 0.5735 0.2353 0.1029 ...
  ..$ root.edge  : num 0
  ..- attr(*, "class")= chr "phylo"
  ..- attr(*, "order")= chr "cladewise"
}}
\usage{
eklof
}
\description{
The object 'eklof' has two items: a 'network' object and a 'tree' object.
The 'network' object is a binary SxS matrix where a value of 1 indicates
that the row species consumes the column species. The 'tree' object is a
phylogenetic tree of class 'phylo' (see 'ape::read.tree'). The data are a
synthetic dataset created to demonstrate how to use the 'phyloint' package
and therefore do not have an associated reference.
}
\keyword{datasets}

