% Generated by roxygen2 (4.0.1): do not edit by hand
\name{minimize.phylo.gaps}
\alias{minimize.phylo.gaps}
\title{Minimize diet gaps in an adjacency matrix based on prey phylogeny}
\usage{
minimize.phylo.gaps(adj, phy, t.min = 1e-05, t.cooling = 0.99,
  swap.factor = 1, verbose = FALSE)
}
\arguments{
\item{adj}{an adjacency matrix of class 'matrix'}

\item{phy}{a phylogenetic tree of class 'phylo'}
}
\description{
Minimize diet gaps in an adjacency matrix based on prey phylogeny
}

