% Generated by roxygen2 (4.0.1): do not edit by hand
\name{minimize.phylo.gaps.sa}
\alias{minimize.phylo.gaps.sa}
\title{Minimize diet gaps in an adjacency matrix based on prey phylogeny using simulated annealing}
\usage{
minimize.phylo.gaps.sa(adj, phy, t.start = NA, p.start = 0.5,
  start.reps = 100, start.type = "mean", n.temps = NA, t.min = 1e-05,
  t.cooling = 0.99, swap.factor = 1, verbose = FALSE)
}
\arguments{
\item{adj}{an adjacency matrix of class 'matrix'}

\item{phy}{a phylogenetic tree of class 'phylo'}
}
\description{
Minimize diet gaps in an adjacency matrix based on prey phylogeny using simulated annealing
}

