% Generated by roxygen2 (4.0.2): do not edit by hand
\name{minimize.phylo.gaps.sd}
\alias{minimize.phylo.gaps.sd}
\title{Minimize diet gaps in an adjacency matrix based on prey phylogeny using steepest descent}
\usage{
minimize.phylo.gaps.sd(adj, phy, verbose = FALSE)
}
\arguments{
\item{adj}{an adjacency matrix of class 'matrix'}

\item{phy}{a phylogenetic tree of class 'phylo'}
}
\description{
Minimize diet gaps in an adjacency matrix based on prey phylogeny using steepest descent
}

