% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/overlap.combinations.R
\name{overlap.combinations}
\alias{overlap.combinations}
\title{Calculate the overlap based on combinations of traits}
\usage{
overlap.combinations(adj, omat)
}
\arguments{
\item{adj}{The adjacency matrix}

\item{omat}{The overlap matrix for the different traits}
}
\value{
The function returns a table with the overlap produced for all possible combinations
of the traits analyzed.
}
\description{
Calculate the overlap based on combinations of traits
}
\examples{
combs <- overlap.combinations(eklof$network, overlap.matrices)
}
\references{
A Eklöf, U Jacob, J Kopp, J Bosch, R Castro-Urgal, NP Chacoff,
   B Dalsgaard, C de Sassi, M Galetti, PR Guimarães, SB Lomáscolo, AM Martín
   González, MA Pizo, R Rader, A Rodrigo, JM Tylianakis, DP Vázquez, and S
   Allesina (2013) The dimensionality of ecological networks. Ecology
   Letters 16:577-583 (doi: 10.1111/ele.12081)
}

