% Generated by roxygen2 (4.0.1): do not edit by hand
\name{phylo.correction}
\alias{phylo.correction}
\title{Remove potential phylogenetic signal from a set of continuous or discrete traits}
\usage{
phylo.correction(phy, traits)
}
\arguments{
\item{phy}{a phylogenetic tree of class 'phylo'}

\item{traits}{a data frame with species as rows and traits as columns}
}
\description{
Remove potential phylogenetic signal from a set of continuous or discrete traits
}
\references{
MA Butler, TW Schoener, and JB Losos (2000) The relationship between
   sexual size dimorphism and habitat use in Greater Antillean Anolis
   lizards. Evolution 54:259-272
}

