% Generated by roxygen2 (4.0.1): do not edit by hand
\name{phylo.signal}
\alias{phylo.signal}
\title{Function to compute phylogenetic signal of a discrete trait based on comparisons of Pagel's lambda}
\usage{
phylo.signal(trait, phy, rep = 999)
}
\arguments{
\item{trait}{vector with same order as phy$tip.label}

\item{phy}{a phylogenetic tree of class 'phylo'}
}
\description{
Function to compute phylogenetic signal of a discrete trait based on comparisons of Pagel's lambda
}
\references{
M Pagel (1999) Inferring the historical patterns of biological
  evolution. Nature 401:877-884
}

