% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sum.phylo.gaps}
\alias{sum.phylo.gaps}
\title{Calculate the number of diet gaps in an adjacency matrix given an ordering compatible with the phylogeny}
\usage{
\method{sum}{phylo.gaps}(adj, phy)
}
\arguments{
\item{adj}{The adjacency matrix}

\item{tc}{The tree coordinate object}
}
\description{
Calculate the number of diet gaps in an adjacency matrix given an ordering compatible with the phylogeny
}

