% Generated by roxygen2 (4.0.1): do not edit by hand
\name{swap.tree}
\alias{swap.tree}
\title{Permute a random pair of sister clades in a phylogenetic tree}
\usage{
swap.tree(phy)
}
\arguments{
\item{phy}{The phylo object}
}
\description{
Permute a random pair of sister clades in a phylogenetic tree
}

