% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tree.coord}
\alias{tree.coord}
\title{Calculate the xy-coordinates of nodes in a phylogenetic tree}
\usage{
tree.coord(phy, type = c("all", "tips"))
}
\arguments{
\item{phy}{The phylo object}

\item{type}{Specify whether to return coordinates for all nodes or just tips}
}
\description{
Calculate the xy-coordinates of nodes in a phylogenetic tree
}

