% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/build.matrix.from.traits.R
\name{build.matrix.from.traits}
\alias{build.matrix.from.traits}
\title{Build a predicted adjacency matrix based on trait values}
\usage{
build.matrix.from.traits(adj, traits, direction = 1)
}
\arguments{
\item{adj}{The adjacency matrix}

\item{traits}{A vector or matrix of continuous trait values}

\item{direction}{Whether to build based on rows (1) or columns (2)}
}
\description{
Build a predicted adjacency matrix based on trait values
}
\references{
A Eklöf, U Jacob, J Kopp, J Bosch, R Castro-Urgal, NP Chacoff,
   B Dalsgaard, C de Sassi, M Galetti, PR Guimarães, SB Lomáscolo, AM Martín
   González, MA Pizo, R Rader, A Rodrigo, JM Tylianakis, DP Vázquez, and S
   Allesina (2013) The dimensionality of ecological networks. Ecology
   Letters 16:577-583 (doi: 10.1111/ele.12081)
}

