% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phylo.correction.R
\name{phylo.correction}
\alias{phylo.correction}
\title{Remove potential phylogenetic signal from a set of continuous or discrete traits}
\usage{
phylo.correction(traits, phy)
}
\arguments{
\item{traits}{A data frame with species as rows and traits as columns}

\item{phy}{A phylogenetic tree of class 'phylo'}
}
\description{
Remove potential phylogenetic signal from a set of continuous or discrete traits
}
\references{
MA Butler, TW Schoener, and JB Losos (2000) The relationship between
   sexual size dimorphism and habitat use in Greater Antillean Anolis
   lizards. Evolution 54:259-272 (doi: 10.1111/j.0014-3820.2000.tb00026.x)
}

