% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.phylo.gaps.R
\name{plot.phylo.gaps}
\alias{plot.phylo.gaps}
\title{Plot a tree coordinate object in the xy plane and an ordered adjacency matrix by its side}
\usage{
\method{plot}{phylo.gaps}(x, type = "optim", tree.lwd = 1, link.size = 1)
}
\arguments{
\item{x}{an object of class 'phylo.gaps'}
}
\description{
Plot a tree coordinate object in the xy plane and an ordered adjacency matrix by its side
}

