% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.phylo.gaps}
\alias{plot.phylo.gaps}
\title{Plot a tree coordinate object in the xy plane and an ordered adjacency matrix by its side}
\usage{
\method{plot}{phylo.gaps}(x, tree.lwd = 1, link.size = 1)
}
\arguments{
\item{x}{an object of class 'phylo.gaps'}
}
\description{
Plot a tree coordinate object in the xy plane and an ordered adjacency matrix by its side
}

