% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sa.start}
\alias{sa.start}
\title{Estimate a starting temperature for simulated annealing}
\usage{
sa.start(adj, phy, prob = 0.5, reps = 100)
}
\arguments{
\item{adj}{The adjacency matrix}

\item{phy}{The phylo object}

\item{prob}{The target initial probability of accepting a move; defaults to 0.50}

\item{reps}{The number of reps to estimate across; defaults to 100}
}
\description{
Estimate a starting temperature for simulated annealing
}

