% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sum.phylo.gaps.R
\name{sum.phylo.gaps}
\alias{sum.phylo.gaps}
\title{Calculate the number of diet gaps in an adjacency matrix given an ordering compatible with the phylogeny}
\usage{
\method{sum}{phylo.gaps}(adj, phy)
}
\arguments{
\item{adj}{The adjacency matrix}

\item{phy}{A phylogenetic tree of class 'phylo'}
}
\description{
Calculate the number of diet gaps in an adjacency matrix given an ordering compatible with the phylogeny
}

