% Generated by roxygen2 (4.0.1): do not edit by hand
\name{swap.tree}
\alias{swap.tree}
\title{Permute a set pair of sister clades in a phylogenetic tree}
\usage{
swap.tree(phy, node = NULL)
}
\arguments{
\item{phy}{The phylo object}

\item{node}{The node to permuate around}
}
\description{
Permute a set pair of sister clades in a phylogenetic tree
}

