% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/swap.tree.R
\name{swap.tree}
\alias{swap.tree}
\title{Permute a pair of sister clades in a phylogenetic tree}
\usage{
swap.tree(phy, node = NULL)
}
\arguments{
\item{phy}{A phylogenetic tree of class 'phylo'}

\item{node}{The node to permutate around}
}
\description{
Permute a pair of sister clades in a phylogenetic tree
}

