% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tree.coord.R
\name{tree.coord}
\alias{tree.coord}
\title{Calculate the xy-coordinates of nodes in a phylogenetic tree}
\usage{
tree.coord(phy, type = c("all", "tips"))
}
\arguments{
\item{phy}{A phylogenetic tree of class 'phylo'}

\item{type}{Specify whether to return coordinates for all nodes or just tips}
}
\description{
Calculate the xy-coordinates of nodes in a phylogenetic tree
}

