% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tree.coord}
\alias{tree.coord}
\title{Calculate the xy-coordinates of nodes in a phylogenetic tree}
\usage{
tree.coord(phy)
}
\arguments{
\item{phy}{The phylo object}
}
\description{
Calculate the xy-coordinates of nodes in a phylogenetic tree
}

